Fix PostgreSQL type mismatch in market price snapshots
- Changed SUM(quantity) result type from i32 to i64 to match PostgreSQL INT8
- Fixed similar issue in death_service major holdings query
- Removed unused imports in market_service
- Server now successfully records price snapshots without errors

+++ src/services/death_service.rs
@@ -184,7 +184,7 @@ impl DeathService {
         }
 
         // Get major holdings for the character
-        let major_holdings: Vec<(Uuid, i32)> = sqlx::query_as(
+        let major_holdings: Vec<(Uuid, i64)> = sqlx::query_as(
             r#"
             SELECT item_id, SUM(quantity) as total_quantity
             FROM character_inventory
